fixed ordering bug in trigram match function
discovered that switching the order of from, to caused the matching
routine to give different results.  Was happening because of the
not-quite-matches of the road, from road, to road names, and the fact
that I was starting building the road always from the "from" approximate
match.

The fix ranks in a UNION statement the from matches and the to matches,
picking the highest ranked match to use as the "from" node and to start
building the road.